Fix config resetting
# Conflicts:
#	src/main/java/com/bobmowzie/mowziesmobs/server/ai/animation/AnimationFWNAttackAI.java
#	src/main/java/com/bobmowzie/mowziesmobs/server/ai/animation/AnimationFWNStompAttackAI.java
#	src/main/java/com/bobmowzie/mowziesmobs/server/ai/animation/AnimationFWNVerticalAttackAI.java
#	src/main/java/com/bobmowzie/mowziesmobs/server/entity/effects/EntityAxeAttack.java

+++ src/main/java/com/bobmowzie/mowziesmobs/server/entity/barakoa/EntityBarako.java
@@ -197,7 +197,7 @@ public class EntityBarako extends MowzieEntity implements LeaderSunstrikeImmune,
                 }
             }
         });
-        this.goalSelector.addGoal(2, new AnimationRadiusAttack<EntityBarako>(this, ATTACK_ANIMATION, 4f, ConfigHandler.COMMON.MOBS.BARAKO.combatConfig.attackMultiplier.get(), 3f, 12, true){
+        this.goalSelector.addGoal(2, new AnimationRadiusAttack<EntityBarako>(this, ATTACK_ANIMATION, 4f, ConfigHandler.COMMON.MOBS.BARAKO.combatConfig.attackMultiplier.get().floatValue(), 3f, 12, true){
             @Override
             public void startExecuting() {
                 super.startExecuting();
